requirements.txt
anthropic==0.40.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
schedule==1.2.1
lxml==5.1.0
